+++ examples/platform/silabs/efr32/wf200/efr_spi.c
@@ -40,7 +40,6 @@
 #include "spidrv.h"
 
 #include "spi_multiplex.h"
-#include "wfx_host_events.h"
 
 #if defined(SL_CATALOG_POWER_MANAGER_PRESENT)
 #include "sl_power_manager.h"
