+++ src/platform/silabs/BLEManagerImpl.h
@@ -27,13 +27,7 @@
 #include "FreeRTOS.h"
 #include "timers.h"
 #if (SLI_SI91X_ENABLE_BLE || RSI_BLE_ENABLE)
-#ifdef __cplusplus
-extern "C" {
-#endif // __cplusplus
 #include "wfx_sl_ble_init.h"
-#ifdef __cplusplus
-}
-#endif // __cplusplus
 #else
 #include "gatt_db.h"
 #include "sl_bgapi.h"
