+++ src/platform/silabs/rs911x/rsi_ble_config.h
@@ -16,12 +16,20 @@
  ******************************************************************************/
 #pragma once
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 #include "rsi_ble_apis.h"
 #if (SLI_SI91X_MCU_INTERFACE | EXP_BOARD)
 #include "rsi_bt_common_apis.h"
 #include "rsi_user.h"
 #else
 #include <rsi_data_types.h>
+#endif
+
+#ifdef __cplusplus
+}
 #endif
 
 #if SL_MATTER_GN_BUILD == 0
