+++ src/platform/silabs/rs911x/wfx_sl_ble_init.cpp
@@ -20,9 +20,11 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
+
 #include <lib/support/logging/CHIPLogging.h>
 #include <platform/CHIPDeviceLayer.h>
 #include <platform/silabs/BLEManagerImpl.h>
+#include <platform/silabs/rs911x/wfx_sl_ble_init.h>
 
 using namespace chip::DeviceLayer::Internal;
